+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project follows versions of format `{year}.{month}.{patch_number}`.
 
 - Updated dependencies to match those in tergite-backend e.g. "quantify-scheduler~=0.22.2"
 - Removed the `LOG_DIR` option on SessionContext so that all that is needed is the `DATA_DIR`
+- Changed the `FIXED_DURATION_QUBITS` SessionContext option to `FIXED_DURATION_COUPLERS` since it is the couplers
+  that would have the fixed duration, not the qubits.
 
 ## [2026.05.0] - 2026-05-09